+++ src/HcpManager.cpp
@@ -147,9 +147,9 @@ std::vector<std::string> HcpManager::GetAllSensorNames()
     return sensorNames;
 }
 
-std::map<std::string, int32> HcpManager::GetAllSensors()
+std::map<std::string, double> HcpManager::GetAllSensors()
 {
-    map<std::string, int32> sensors;
+    map<std::string, double> sensors;
 
     json analogs = mapping->GetDevices(Device_Type::ANALOG);
     json digitals = mapping->GetDevices(Device_Type::DIGITAL);
@@ -737,9 +737,9 @@ int32 *HcpManager::GetLoadCells()
     return value;
 }
 
-int32 HcpManager::GetAnalog(std::string name)
+double HcpManager::GetAnalog(std::string name)
 {
-    int32 value = -1;
+    double value = -1;
 
     json device = mapping->GetDeviceByName(name, Device_Type::ANALOG);
 
@@ -772,7 +772,7 @@ int32 HcpManager::GetAnalog(std::string name)
 
             json map = device["map"];
 
-            int32 before = value;
+            double before = value;
             value = ((double)value - (double)map["d"]) * (double)map["k"];
 
             Debug::info("from %d to %d", before, value);
