+++ src/LLInterface.cpp
@@ -50,10 +50,10 @@ std::vector<std::string> LLInterface::GetAllSensorNames()
     return sensorNames;
 }
 
-std::map<std::string, int32> LLInterface::GetAllSensors()
+std::map<std::string, double> LLInterface::GetAllSensors()
 {
     auto startTime = Clock::now();
-    std::map<std::string, int32> sensors;
+    std::map<std::string, double> sensors;
     sensors = HcpManager::GetAllSensors();
 
     auto currTime = Clock::now();
@@ -102,7 +102,7 @@ void LLInterface::StopGetSensors()
 
 void LLInterface::GetSensors(int64 microTime)
 {
-    std::map<std::string, int32> sensors = GetAllSensors();
+    std::map<std::string, double> sensors = GetAllSensors();
 
     TransmitSensors(microTime, sensors);
 
@@ -130,7 +130,7 @@ void LLInterface::GetSensors(int64 microTime)
 
 }
 
-void LLInterface::TransmitSensors(int64 microTime, std::map<std::string, int32> sensors)
+void LLInterface::TransmitSensors(int64 microTime, std::map<std::string, double> sensors)
 {
     json content = json::array();
     json sen;
